usenostr hook pulled in yay
